compilation errors
display and keyboard drivers may need to store more state, which would also require the function pointers to take in a driver pointer. This would be a royal pain in the ass. To make the current design work, I would have to figure out where to init the keyboard driver when the terminal is already set, or a have a function outside of the kd struct to set the terminal, which is bad design (at least im 99% sure)